refactor(query): collapse the lazy function dispatcher onto the eager registry (dual-eval PR 5)
evaluate_function now delegates every non-row-context function to
evaluate_function_on_values (keeping only META-family, COALESCE, and the
aggregate per-row stubs on the lazy path). Reconcile 6 latent eager-registry
bugs the collapse exposed (MAXWIDTH naive-truncation, ROUND negative-precision
no-op, JOINSTR separator, POSSIGN integer arm, LENGTH on sets, CONVERT
amount-string), unifying on the tested Python-correct lazy behavior via shared
value-cores. Delete 28 now-dead eval_* helpers + the emptied position module
(~1620 lines). One implementation per function. Architecture roadmap [XL/BR9].
