+++ crates/rustledger-query/src/executor/functions/mod.rs
@@ -5,7 +5,6 @@
 mod account;
 mod date;
 mod math;
-mod position;
 pub(in crate::executor) mod string;
 mod util;
