+++ crates/rustledger-query/src/executor/functions/util.rs
@@ -3,7 +3,7 @@
 //! This module includes metadata, conversion, casting, and helper functions.
 
 use rust_decimal::Decimal;
-use rustledger_core::{Amount, Inventory, MetaValue, Metadata, Position};
+use rustledger_core::{MetaValue, Metadata};
 
 use crate::ast::FunctionCall;
 use crate::error::QueryError;
@@ -122,213 +122,6 @@ impl Executor<'_> {
         }
     }
 
-    /// Evaluate CONVERT function (currency conversion).
-    ///
-    /// `CONVERT(position, currency)` - Convert position/amount to target currency.
-    /// `CONVERT(position, currency, date)` - Convert using price at specific date.
-    pub(crate) fn eval_convert(
-        &self,
-        func: &FunctionCall,
-        ctx: &PostingContext,
-    ) -> Result<Value, QueryError> {
-        if func.args.len() < 2 || func.args.len() > 3 {
-            return Err(QueryError::InvalidArguments(
-                "CONVERT".to_string(),
-                "expected 2 or 3 arguments: (value, currency[, date])".to_string(),
-            ));
-        }
-
-        let val = self.evaluate_expr(&func.args[0], ctx)?;
-
-        let target_currency = match self.evaluate_expr(&func.args[1], ctx)? {
-            Value::String(s) => s,
-            Value::Null => {
-                return Err(QueryError::Type(
-                    concat!(
-                        "CONVERT: second argument evaluated to NULL; ",
-                        "expected a currency string ",
-                        "(this often means an aggregate expression couldn't ",
-                        "evaluate against an empty group — see issue #902)",
-                    )
-                    .to_string(),
-                ));
-            }
-            _ => {
-                return Err(QueryError::Type(
-                    "CONVERT: second argument must be a currency string".to_string(),
-                ));
-            }
-        };
-
-        // When no date is specified, use the latest price (matches Python beancount behavior)
-        let date: Option<rustledger_core::NaiveDate> = if func.args.len() == 3 {
-            match self.evaluate_expr(&func.args[2], ctx)? {
-                Value::Date(d) => Some(d),
-                _ => {
-                    return Err(QueryError::Type(
-                        "CONVERT: third argument must be a date".to_string(),
-                    ));
-                }
-            }
-        } else {
-            None // Use latest price when no date specified
-        };
-
-        // Helper to convert an amount, using latest price if no date specified
-        let convert_amount = |amt: &Amount| -> Option<Amount> {
-            if let Some(d) = date {
-                self.price_db.convert(amt, &target_currency, d)
-            } else {
-                self.price_db.convert_latest(amt, &target_currency)
-            }
-        };
-
-        match val {
-            Value::Position(p) => {
-                if p.units.currency == target_currency {
-                    Ok(Value::Amount(p.units))
-                } else if let Some(converted) = convert_amount(&p.units) {
-                    Ok(Value::Amount(converted))
-                } else {
-                    // Return original units if no conversion available
-                    Ok(Value::Amount(p.units))
-                }
-            }
-            Value::Amount(a) => {
-                if a.currency == target_currency {
-                    Ok(Value::Amount(a))
-                } else if let Some(converted) = convert_amount(&a) {
-                    Ok(Value::Amount(converted))
-                } else {
-                    Ok(Value::Amount(a))
-                }
-            }
-            Value::Inventory(inv) => {
-                // Convert each position, keeping originals when no conversion available
-                // (matches Python beancount behavior)
-                let mut result = Inventory::default();
-                for pos in inv.positions() {
-                    if pos.units.currency == target_currency {
-                        result.add(Position::simple(pos.units.clone()));
-                    } else if let Some(converted) = convert_amount(&pos.units) {
-                        result.add(Position::simple(converted));
-                    } else {
-                        // No conversion available - keep original (Python beancount behavior)
-                        result.add(Position::simple(pos.units.clone()));
-                    }
-                }
-                // If result has single currency matching target, return as Amount.
-                // If result is empty, return zero in target currency (issue #586).
-                // Peek the first two positions to decide; the single-match Amount
-                // is cloned eagerly (O(1) — `Decimal` + interned currency) so the
-                // iterator borrow is dropped before we move `result` below.
-                let single_match: Option<Amount> = {
-                    let mut iter = result.positions();
-                    match (iter.next(), iter.next()) {
-                        (Some(only), None) if only.units.currency == target_currency => {
-                            Some(only.units.clone())
-                        }
-                        _ => None,
-                    }
-                };
-                if let Some(units) = single_match {
-                    Ok(Value::Amount(units))
-                } else if result.is_empty() {
-                    Ok(Value::Amount(Amount::new(Decimal::ZERO, &target_currency)))
-                } else {
-                    Ok(Value::Inventory(Box::new(result)))
-                }
-            }
-            Value::Number(n) => {
-                // Just wrap the number as an amount with the target
-                // currency. Note the asymmetry with the `Value::String`
-                // arm below: `CONVERT(100, 'EUR')` returns `100 EUR`
-                // (no conversion — the bare number has no source
-                // currency), but `CONVERT('100 USD', 'EUR')` does an
-                // actual conversion.
-                Ok(Value::Amount(Amount::new(n, &target_currency)))
-            }
-            Value::String(s) => {
-                // String input is a rustledger extension (issue #1179),
-                // not present in Python beancount. Lets users write
-                // ad-hoc currency conversions like
-                // `SELECT CONVERT('100 USD', 'EUR')` without anchoring
-                // them to a posting. Strict parser (see
-                // `Amount::from_str`): malformed input surfaces as a
-                // typed `QueryError` rather than a silent zero or a
-                // panic.
-                let amt: Amount = s.parse().map_err(|e| {
-                    QueryError::Type(format!("CONVERT: first argument {e} (e.g. \"100 USD\")"))
-                })?;
-                if amt.currency == target_currency {
-                    Ok(Value::Amount(amt))
-                } else if let Some(converted) = convert_amount(&amt) {
-                    Ok(Value::Amount(converted))
-                } else {
-                    // Match the `Value::Amount` arm: no price available
-                    // → return original unchanged. Switching to an
-                    // error here would diverge from
-                    // `CONVERT(<amount>, ...)` for the same
-                    // "no rate found" case.
-                    Ok(Value::Amount(amt))
-                }
-            }
-            Value::Null => {
-                // For null values (e.g., empty sum), return zero in target currency
-                // This matches Python beancount behavior for empty balances
-                Ok(Value::Amount(Amount::new(Decimal::ZERO, &target_currency)))
-            }
-            _ => Err(QueryError::Type(
-                "CONVERT expects a position, amount, inventory, number, or amount-string"
-                    .to_string(),
-            )),
-        }
-    }
-
-    /// Evaluate INT function (convert to integer).
-    pub(crate) fn eval_int(
-        &self,
-        func: &FunctionCall,
-        ctx: &PostingContext,
-    ) -> Result<Value, QueryError> {
-        Self::require_args("INT", func, 1)?;
-        let val = self.evaluate_expr(&func.args[0], ctx)?;
-        Self::value_to_int(&val)
-    }
-
-    /// Evaluate DECIMAL function (convert to decimal).
-    pub(crate) fn eval_decimal(
-        &self,
-        func: &FunctionCall,
-        ctx: &PostingContext,
-    ) -> Result<Value, QueryError> {
-        Self::require_args("DECIMAL", func, 1)?;
-        let val = self.evaluate_expr(&func.args[0], ctx)?;
-        Self::value_to_decimal(&val)
-    }
-
-    /// Evaluate STR function (convert to string).
-    pub(crate) fn eval_str(
-        &self,
-        func: &FunctionCall,
-        ctx: &PostingContext,
-    ) -> Result<Value, QueryError> {
-        Self::require_args("STR", func, 1)?;
-        let val = self.evaluate_expr(&func.args[0], ctx)?;
-        Self::value_to_str(&val)
-    }
-
-    /// Evaluate BOOL function (convert to boolean).
-    pub(crate) fn eval_bool(
-        &self,
-        func: &FunctionCall,
-        ctx: &PostingContext,
-    ) -> Result<Value, QueryError> {
-        Self::require_args("BOOL", func, 1)?;
-        let val = self.evaluate_expr(&func.args[0], ctx)?;
-        Self::value_to_bool(&val)
-    }
-
     // =========================================================================
     // Value conversion helpers (shared between eval_* and evaluate_function_on_values)
     // =========================================================================
@@ -425,64 +218,5 @@ impl Executor<'_> {
             }
         }
         Ok(Value::Null)
-    }
-
-    /// Evaluate ONLY function.
-    ///
-    /// `ONLY(key, inventory)` - Extract amount with given currency from inventory.
-    /// Returns the amount if exactly one position matches, NULL otherwise.
-    pub(crate) fn eval_only(
-        &self,
-        func: &FunctionCall,
-        ctx: &PostingContext,
-    ) -> Result<Value, QueryError> {
-        Self::require_args("ONLY", func, 2)?;
-
-        // Get the currency key
-        let key = match self.evaluate_expr(&func.args[0], ctx)? {
-            Value::String(s) => s,
-            _ => {
-                return Err(QueryError::Type(
-                    "ONLY: first argument must be a currency string".to_string(),
-                ));
-            }
-        };
-
-        // Get the inventory
-        let inv = match self.evaluate_expr(&func.args[1], ctx)? {
-            Value::Inventory(inv) => inv,
-            Value::Position(pos) => {
-                // If it's a single position, check if it matches
-                if pos.units.currency == key {
-                    return Ok(Value::Amount(pos.units));
-                }
-                return Ok(Value::Null);
-            }
-            Value::Amount(amt) => {
-                // If it's a single amount, check if it matches
-                if amt.currency == key {
-                    return Ok(Value::Amount(amt));
-                }
-                return Ok(Value::Null);
-            }
-            Value::Null => return Ok(Value::Null),
-            _ => {
-                return Err(QueryError::Type(
-                    "ONLY: second argument must be an inventory, position, or amount".to_string(),
-                ));
-            }
-        };
-
-        // Find positions matching the currency
-        let matching: Vec<_> = inv
-            .positions()
-            .filter(|p| p.units.currency == key)
-            .collect();
-
-        match matching.len() {
-            0 => Ok(Value::Null),
-            1 => Ok(Value::Amount(matching[0].units.clone())),
-            _ => Ok(Value::Null), // Multiple matches, return NULL
-        }
     }
 }
